requirements.txt
attr==0.3.2
brotli==1.0.9
brotlicffi==1.0.9.2
ConfigParser==5.3.0
contextlib2==21.6.0
cryptography==40.0.2
Cython==0.29.34
dl==0.1.0
docutils==0.20.1
HTMLParser==0.0.2
importlib_metadata==6.6.0
ipython==8.13.2
ipywidgets==8.0.6
Jinja2==3.1.2
jnius==1.1.0
keyring==23.13.1
langchain==0.0.173
lockfile==0.12.2
mock==5.0.2
numpy==1.24.3
ordereddict==1.1
Pillow==9.5.0
Pillow==9.5.0
protobuf==3.20.3
pyOpenSSL==23.1.1
python-dotenv==1.0.0
railroad==0.5.0
simplejson==3.19.1
Sphinx==7.0.1
streamlit==1.22.0
streamlit_chat==0.0.2.2
tornado==6.3.2
truststore==0.7.0
urllib3_secure_extra==0.1.0
wincertstore==0.2.1
xmlrpclib==1.0.1
zstandard==0.21.0



# apt install -y make build-essential libssl-dev zlib1g-dev \libbz2-dev libreadline-dev libsqlite3-dev wget curl llvm libncurses5-dev\libncursesw5-dev xz-utils tk-dev libffi-dev liblzma-dev python-openssl\git